Remove identifier and link the title to source db
