+++ src/uniparc/components/sub-entry/SubEntryOverview.tsx
@@ -1,53 +1,16 @@
-import { ExternalLink, InfoList, Loader } from 'franklin-sites';
+import { InfoList } from 'franklin-sites';
 import { memo } from 'react';
 import { Link } from 'react-router-dom';
 
 import TaxonomyView from '../../../shared/components/entry/TaxonomyView';
-import apiUrls from '../../../shared/config/apiUrls/apiUrls';
-import useDataApi from '../../../shared/hooks/useDataApi';
-import { Namespace } from '../../../shared/types/namespaces';
-import { type UniParcXRef } from '../../adapters/uniParcConverter';
 import { type UniParcSubEntryUIModel } from '../../adapters/uniParcSubEntryConverter';
 import EntrySection from '../../types/subEntrySection';
-import { type DataDBModel } from '../entry/XRefsSection';
-
-type ExternalXrefLinkProps = { xref: UniParcXRef; dataDB: DataDBModel };
-
-const ExternalXrefLink = ({ xref, dataDB }: ExternalXrefLinkProps) => {
-  let { id } = xref;
-  if (!id || !xref.database) {
-    return null;
-  }
-  const template = dataDB.find(
-    ({ displayName }) => displayName === xref.database
-  )?.uriLink;
-  if (!template) {
-    return null;
-  }
-  // NOTE: exception for FusionGDB we need to remove the underscore number
-  if (xref.database === 'FusionGDB') {
-    id = id.replace(/_\d+$/, '');
-  }
-  return (
-    <ExternalLink url={template.replace('%id', id)}>
-      {xref.id}
-      {xref.chain && ` (chain ${xref.chain})`}
-    </ExternalLink>
-  );
-};
 
 type Props = {
   data: UniParcSubEntryUIModel;
 };
 
 const SubEntryOverview = ({ data }: Props) => {
-  const dataDB = useDataApi<DataDBModel>(
-    apiUrls.configure.allDatabases(Namespace.uniparc)
-  );
-  if (dataDB.loading || !dataDB.data) {
-    return <Loader />;
-  }
-
   const infoData = [
     {
       title: <span data-article-id="protein_names">Protein</span>,
@@ -59,12 +22,6 @@ const SubEntryOverview = ({ data }: Props) => {
       title: <span data-article-id="gene_name">Gene</span>,
       content: data.subEntry.geneName && (
         <strong>{data.subEntry.geneName}</strong>
-      ),
-    },
-    {
-      title: 'Identifier',
-      content: !data.subEntry.isUniprotkbEntry && (
-        <ExternalXrefLink xref={data.subEntry} dataDB={dataDB.data} />
       ),
     },
     {
